Search page
ok so im fucking losing my mind rn so what im saying here may not make sense. All the search functions are in the /about or about.html shit. the function is in blog/views.py the issue is that when you go to the search and then go back to home it breaks